alex/branching-chat-hosting: remove durable object, handle AI requests directly in worker
Simplifies architecture by eliminating unnecessary DO proxy layer.
Worker now handles OpenAI requests directly for better performance.
