+++ templates/branching-chat/wrangler.toml
@@ -3,13 +3,6 @@ main = "worker/worker.ts"
 compatibility_date = "2024-12-30"
 assets = { not_found_handling = "single-page-application" }
 
-# Set up the durable object used for each tldraw room
-[durable_objects]
-bindings = [
-    { name = "MY_DURABLE_OBJECT", class_name = "MyDurableObject" },
-]
-
-# Durable objects require migrations to create/modify/delete them
-[[migrations]]
-tag = "v1"
-new_classes = ["MyDurableObject"]
+[[routes]]
+pattern = "branching-chat.templates.tldraw.dev"
+custom_domain = true
